requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
PyPDF2>=3.0.1
python-docx>=0.8.11
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.2
requests>=2.31.0